Allow switching sign from false <=> true
sign can either force a new resource to be created or the resource must
support update. If it is not the case terraform will destroy/apply but
this is source of issue for puppet as certificate is only requested
once. If terraform destroy the certificate it will not be able to apply
afterward. Defining an update method allow to not trigger this behavior.
The easiest way to do it is to set the update method to
resourcePuppetCACertificateCreate which is idempotent